layout working but compilcated between clear and draw border box

+++ la.c
@@ -1,7 +1,7 @@
 #include "la.h"
 
 Vec2i
-vec2i(int x, int y)
+vec2i(int16_t x, int16_t y)
 {
     Vec2i newVec;
     newVec.x = x;
